requirements.txt
fastapi
sqlalchemy
strawberry-graphql
pytest
uvicorn[standard]
httpx